examples/voice-pipeline-agent/requirements.txt
livekit-agents>=0.12.17
livekit-plugins-deepgram>=0.7.1
livekit-plugins-google>=0.11.1
livekit-plugins-openai[vertex]>=0.10.10,<1.0.0
livekit-plugins-silero>=0.7.5
livekit-plugins-rag>=0.2.4
python-dotenv~=1.0
aiofile~=3.8.8
g2p-en
